Tests/LibWeb/Text/expected/css/accent-color-keyword.txt

AccentColor with custom accent-color: rgb(255, 0, 0)
PASS: AccentColor respects element's accent-color property
